backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
redis==5.0.4
rq==1.16.2
pydantic==2.7.4
python-multipart==0.0.9
pdfminer.six==20240706
